requirements.txt
openai-clip
torch
torchvision
pillow
pinecone
tqdm
langgraph
tensorflow.keras
tavily
langchain-openai
langchain-core
langchain
requests
selenium
beautifulsoup4